project/server/requirements.txt
appdirs==1.4.3
appnope==0.1.0
argcomplete==1.12.3
argon2-cffi==21.1.0
asgiref==3.4.1
astroid==2.2.5
attrs==21.2.0
backcall==0.1.0
bleach==4.1.0
boto3==1.12.47
botocore==1.15.47
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.3
cycler==0.10.0
debugpy==1.4.3
decorator==4.4.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.0.3
docutils==0.15.2
entrypoints==0.3
fastapi==0.68.1
filelock==3.0.12
h11==0.12.0
idna==3.2
importlib-metadata==1.5.0
iota==1.4.20
iota-client-python==0.2.0a26
iota-wallet==0.1.0a2
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==4.3.21
jedi==0.17.0
Jinja2==3.0.2
jmespath==0.9.5
jsonschema==4.1.0
jupyter==1.0.0
jupyter-client==7.0.5
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
lazy-object-proxy==1.4.2
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.4
numpy==1.21.2
packaging==21.0
pandas==1.3.3
pandocfilters==1.5.0
parso==0.7.0
pexpect==4.8.0
phx-class-registry==3.0.5
phx-filters==2.0.2
pickleshare==0.7.5
Pillow==8.3.2
prometheus-client==0.11.0
prompt-toolkit==3.0.5
psycopg2==2.7.5
ptyprocess==0.6.0
pycparser==2.20
pydantic==1.8.2
Pygments==2.6.1
pylint==2.3.1
PyOTA==2.1.0
PyOTA-CCurl==1.0.9
PyOTA-PoW==1.1.0
pyparsing==2.4.7
pyrsistent==0.18.0
pysha3==1.0.2
python-dateutil==2.8.1
python-dotenv==0.19.0
pytz==2018.9
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
regex==2021.9.24
requests==2.26.0
s3transfer==0.3.3
Send2Trash==1.8.0
six==1.12.0
sqlparse==0.3.0
starlette==0.14.2
terminado==0.12.1
testpath==0.5.0
tornado==6.1
traitlets==5.1.0
typed-ast==1.4.0
typing-extensions==3.10.0.2
urllib3==1.25.9
uvicorn==0.15.0
virtualenv==20.0.1
wcwidth==0.1.9
webencodings==0.5.1
whitenoise==4.1.2
widgetsnbextension==3.5.1
wrapt==1.11.2
wxPython==4.1.1
zipp==2.2.0
